read study title and descr from study + remove unused project xml generator
