+++ test-data/biosamples-input-isa-multi.json
@@ -56,8 +56,8 @@
     "studies": [
       {
         "identifier": "study1",
-        "title": "Arabidopsis thaliana",
-        "description": "Nucleic acid sequencing and metabolomics and proteomics of Arabidopsis thaliana in specific experimental conditions to test a specific hypothesis.\r\n",
+        "title": "Integrated multi-omics profiling of Arabidopsis thaliana under controlled experimental conditions",
+        "description": "A coordinated multi-omics study combining sequencing-based and molecular profiling approaches to characterize Arabidopsis thaliana responses under controlled experimental conditions.\r\n",
         "submissionDate": "",
         "publicReleaseDate": "",
         "filename": "s_Arabidopsis thaliana.txt",
@@ -829,16 +829,8 @@
                 "value": "ena"
               },
               {
-                "name": "STUDY_TITLE",
-                "value": "Arabidopsis thaliana sequencing project"
-              },
-              {
-                "name": "STUDY_DESCRIPTION",
-                "value": "Nucleic acid sequencing and metabolomics and proteomics of Arabidopsis thaliana in specific experimental conditions to test a specific hypothesis."
-              },
-              {
                 "name": "STUDY_ABSTRACT",
-                "value": "Nucleic acid sequencing and metabolomics and proteomics of Arabidopsis thaliana in specific experimental conditions to test a specific hypothesis."
+                "value": "A coordinated multi-omics study combining sequencing-based and molecular profiling approaches to characterize Arabidopsis thaliana responses under controlled experimental conditions."
               },
               {
                 "name": "STUDY_TYPE",
